requirements.txt
# AI
anthropic>=0.40.0

# Google Workspace
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Microsoft Teams (Graph API)
msal>=1.28.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
